+++ dbparti/management/commands/partition.py
@@ -1,5 +1,5 @@
 from dbparti.models import Partitionable
-from django.db.models import get_models, connection
+from django.db.models import get_models
 from django.core.management.base import AppCommand
 
 
@@ -14,11 +14,8 @@ class Command(AppCommand):
             if issubclass(model, Partitionable):
                 names.append(model.__name__)
 
-                db = getattr(__import__('dbparti.backends.' + connection.vendor,
-                    fromlist=[connection.vendor.capitalize()]), connection.vendor.capitalize())(
-                        model._meta.db_table, model._meta.partition_column)
-
-                db.init_partition(model._meta.partition_range)
+                model_instance = model()
+                model_instance.db.init_partition(model._meta.partition_range)
 
         if not names:
             self.stderr.write('Unable to find any partitionable models in an app: ' + app.__name__.split('.')[0])